fix(frontend): fix sidebar test regression from memory-leak commit
The global afterEach added in 0515fe882 imported both cleanupTestHarness
(from src/test-utils.tsx) and resetAllZustandStores (from
tests/reset-zustand-stores.ts). Both modules transitively load
src/config at setup time — test-utils imports routeTree.gen which
imports every route file, reset-zustand imports backend-api which
imports config directly. That happened BEFORE any test file's
vi.mock('config', ...) hoists could take effect, pinning
isEmbedded/isAdpEnabled live bindings in downstream modules (notably
utils/route-utils.tsx) to the real implementation.

sidebar.test.tsx mocks '../../config' to force isEmbedded() and
isAdpEnabled() to return true so the AGENTIC section (including
/transcripts) becomes visible. After the memory-leak commit, that mock
was a no-op and the test expected /transcripts to be defined but got
undefined.

Fix:
- Extract cleanupTestHarness and its tracking sets into a new
  side-effect-free file frontend/tests/harness-cleanup.ts that only
  imports QueryClient from @tanstack/react-query. vitest.setup.integration.ts
  imports from there instead of src/test-utils.
- Keep resetAllZustandStores synchronous with static imports, but do
  NOT import it from the global setup. Per-file opt-in — a page test
  file that accumulates store state imports the helper directly.
- Drop the global zustand reset; keep cleanup + cleanupTestHarness +
  clearAllMocks + clearAllTimers which are the minimum safe shared
  teardown.

All 87 integration files + 39 unit files pass. Wall time 89s
(vs 155s baseline pre-vitest-upgrade, -42%). sidebar.test.tsx now
passes.
